+++ support/handlers/__init__.py
@@ -5,7 +5,6 @@ from .os2_handler import OS2TableHandler
 from .style_handler import StyleConsistencyHandler
 from .glyph_handler import GlyphHandler
 from .kern_handler import KernHandler
-from .name_handler import NameTableHandler
 
 __all__ = [
     "TableHandler",
@@ -14,5 +13,4 @@ __all__ = [
     "StyleConsistencyHandler",
     "GlyphHandler",
     "KernHandler",
-    "NameTableHandler",
 ]